Refactor data migration to use ResourceConfig
Load base and related resources to RulesFactory facts map used to
compute update values.
