+++ android/engine/src/main/java/org/smartregister/fhircore/engine/util/extension/ResourceExtension.kt
@@ -58,6 +58,7 @@ import org.joda.time.Instant
 import org.json.JSONException
 import org.json.JSONObject
 import org.smartregister.fhircore.engine.data.local.DefaultRepository
+import org.smartregister.fhircore.engine.domain.model.RepositoryResourceData
 import org.smartregister.fhircore.engine.util.fhirpath.FhirPathDataExtractor
 import timber.log.Timber
 
@@ -406,20 +407,20 @@ suspend fun Task.updateDependentTaskDueDate(
  * to be a boolean otherwise the [toBoolean] function will evaluate to false and hence return an
  * empty list.
  */
-fun List<Resource>.filterByFhirPathExpression(
+fun List<RepositoryResourceData>.filterByFhirPathExpression(
   fhirPathDataExtractor: FhirPathDataExtractor,
   conditionalFhirPathExpressions: List<String>?,
   matchAll: Boolean,
-): List<Resource> {
+): List<RepositoryResourceData> {
   if (conditionalFhirPathExpressions.isNullOrEmpty()) return this
-  return this.filter { resource ->
+  return this.filter { repositoryResourceData ->
     if (matchAll) {
       conditionalFhirPathExpressions.all {
-        fhirPathDataExtractor.extractValue(resource, it).toBoolean()
+        fhirPathDataExtractor.extractValue(repositoryResourceData.resource, it).toBoolean()
       }
     } else {
       conditionalFhirPathExpressions.any {
-        fhirPathDataExtractor.extractValue(resource, it).toBoolean()
+        fhirPathDataExtractor.extractValue(repositoryResourceData.resource, it).toBoolean()
       }
     }
   }
